consolidated icon definitions to shared ui
