+++ packages/javascript/bh-shared-ui/src/utils/index.ts
@@ -20,3 +20,4 @@ export * from './exportGraphData';
 export * from './entityInfoDisplay';
 export * from './passwd';
 export * from './user';
+export * from './icons';
